% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOIR.R
\name{LOIR}
\alias{LOIR}
\title{LOIR}
\usage{
LOIR(M0, M1, endogenous, exogenous, moderator, interaction)
}
\arguments{
\item{M0}{the path where mplus output of the model without interaction was saved;}

\item{M1}{the path where mplus output of the model with interaction was saved;}

\item{endogenous}{the endogenous (DV) variable}

\item{exogenous}{the exogenous (IV) variable}

\item{moderator}{the latent moderator variable}

\item{interaction}{the interaction variable created by XWITH in mplus}
}
\description{
R square of interaction between latent and observed variables in 2-way interaction
}
\author{



}
