% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIR.R
\name{LIR}
\alias{LIR}
\title{LIR}
\usage{
LIR(M0, M1, endogenous, exogenous, moderator, interaction)
}
\arguments{
\item{M0}{the Q-matrix specified for the LCDM}

\item{save.path}{save the .stan file to somewhere; the default path is getwd()}

\item{save.name}{name the .stan}
}
\value{
a. stan file saved at the specified path
}
\description{
Latent Variable Interaction
}
\author{

}
