% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIR.R
\name{LIR}
\alias{LIR}
\title{LIR}
\usage{
LIR(M0, M1, endogenous, exogenous, moderator, interaction)
}
\arguments{
\item{M0}{the path where mplus output of model 0 was saved; model 0 is the model without interaction}

\item{M1}{the path where mplus output of model 1 was saved; model 1 is the model with interaction}

\item{endogenous}{the endogenous (DV) variable}

\item{exogenous}{the exogenous (IV) variable}

\item{moderator}{the latent moderator variable}

\item{interaction}{the interaction variable created by XWITH in mplus}
}
\description{
R square of interaction between latent variables in 2-way interaction
}
\author{



}
