% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIR.R
\name{LIR}
\alias{LIR}
\title{LIR}
\usage{
LIR(M0, M1, endogenous, exogenous, moderator, interaction)
}
\arguments{
\item{M0}{the mplus output of the model without interaction;}

\item{M1}{the mplus output of the model with interaction;}

\item{endogenous}{one endogenous latent factor}

\item{exogenous}{one exogenous latent factor}

\item{moderator}{one latent moderator factor}

\item{interaction}{one two-way interaction that are defined by the XWITH function in the mplus}
}
\description{
R square of one two-way interactions between latent factors in structural equation model
}
\author{
{Lu Qin, Howard University, \email{lu.qin@howard.edu}}

{Jihong Zhang, University of Iowa, \email{jihong-zhang@uiowa.edu}}
}
