% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOIR.R
\name{LOIR}
\alias{LOIR}
\title{LOIR}
\usage{
LOIR(M0, M1, endogenous, exogenous, moderator, interaction)
}
\arguments{
\item{M0}{the Q-matrix specified for the LCDM}

\item{M1}{save the .stan file to somewhere; the default path is getwd()}

\item{endogenous}{name the .stan}

\item{exogenous}{name the .stan}

\item{moderator}{name the .stan}

\item{interaction}{name the .stan}
}
\value{
a. stan file saved at the specified path
}
\description{
Latent & Observed Interaction
}
\author{
{Zhehan Jiang, University of Alabama, \email{zjiang17@ua.edu}}
}
