% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOIR.R
\name{LOIR}
\alias{LOIR}
\title{LOIR}
\usage{
LOIR(M0, M1, endogenous, exogenous, moderator, interaction)
}
\arguments{
\item{M0}{the mplus output of the model without interaction;}

\item{M1}{the mplus output of the model with interaction;}

\item{endogenous}{one endogenous latent factor}

\item{exogenous}{one exogenous latent factor}

\item{moderator}{one observed covariate}

\item{interaction}{one two-way interaction that are defined by the XWITH function in the mplus}
}
\description{
R square of one two-way interactions between one latent factor and one observed covariate in structural equation model
}
\author{



}
